worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "openai>=2.30.0",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
]
